+++ src/view/typing.py
@@ -1,11 +1,28 @@
 from __future__ import annotations
 
 from collections.abc import Sequence
-from typing import (TYPE_CHECKING, Any, Awaitable, Callable, Dict, Generic,
-                    List, Literal, Tuple, Type, TypeVar, Union)
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Awaitable,
+    Callable,
+    Dict,
+    Generic,
+    List,
+    Literal,
+    Tuple,
+    Type,
+    TypeVar,
+    Union,
+)
 
-from typing_extensions import (Concatenate, ParamSpec, Protocol, TypedDict,
-                               runtime_checkable)
+from typing_extensions import (
+    Concatenate,
+    ParamSpec,
+    Protocol,
+    TypedDict,
+    runtime_checkable,
+)
 
 if TYPE_CHECKING:
     from _view import Context